ssl_client: change type to uint16_t to store client->available() return value

+++ libraries/SSLClient/src/ssl_client.cpp
@@ -82,11 +82,11 @@ static int client_net_recv_timeout( void *ctx, unsigned char *buf,
     }
     unsigned long start = millis();
     unsigned long tms = start + timeout;
-    int pending = client->available();
+    uint16_t pending = client->available();
     // If there is data in the client, wait for message completion
     if((pending > 0) && (pending < len))
     do {
-        int pending = client->available();
+        uint16_t pending = client->available();
         if (pending < len && timeout > 0) {
             delay(1);
         } else break;
